chore: update index.html to mark page as unmaintained
My homepage is no longer actively maintained, and I have no immediate plans to maintain it. I'm going to keep it up because I want to do something with it eventually (i.e. testbed for projects, AR/VR, micro-frontends, CRMs, backend stuff, you freakin' name it). But, I've got other things to focus on now, so for a while it'll be d-e-d dead.